refactor: adicionando token de autenticacao em todas as rotas de chamada de api

+++ frontend/src/api/http/robos/deleteParametros.ts
@@ -1,6 +1,7 @@
 import { api } from '@/utils/axios';
 
 export async function deleteParametros(roboId: number, parametroId: number) {
-    const res = await api.delete(`robos/${roboId}/parametros/excluir/${parametroId}`).then((res) => res.data);
+    const token = localStorage.getItem('accessToken');
+    const res = await api.delete(`robos/${roboId}/parametros/excluir/${parametroId}`, {headers: {Authorization: `Bearer ${token}`}}).then((res) => res.data);
     return res;
 }
